[31/12/2011] r22 Adding functionalities: - check few bugs - add password age and if password has been changed in registry user's part - review column tri function - add parent key update for MUI and MRU registry's key

+++ RTCA/proc_logs_evtx.c
@@ -139,7 +139,7 @@ unsigned int ReadRecord(char *buffer, DWORD size, STRING_TABLE *my_s_table, LINE
   {
     //lecture des informations du header
     //id
-    snprintf(lv_line[1].c,MAX_LINE_SIZE,"%05lu",h_dheader->RecordNumber);
+    snprintf(lv_line[1].c,MAX_LINE_SIZE,"%05lu",(long unsigned int)(h_dheader->RecordNumber));
 
     FILETIME FileTime, LocalFileTime;
     SYSTEMTIME SysTime;
